requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch
torchvision
torchaudio

sentence-transformers
numpy
faiss-cpu
PyMuPDF
nltk